+++ src/pages/uncontrolled-form-page/ui/UncontrolledFormPage.tsx
@@ -1,6 +1,6 @@
-import { UncontrolledForm } from "@/widgets/uncontrolled-form";
 import { FC } from "react";
 import styles from "./UncontrolledFormPage.module.scss";
+import { UncontrolledForm } from "@/widgets/form";
 
 export const UncontrolledFormPage: FC = () => {
   return (
