--- src/widgets/controlled-form/index.ts
@@ -1,1 +1,0 @@
-export { ControlledForm } from "./ui/ControlledForm";
