+++ src/widgets/form/controlled-form/ui/ControlledForm.module.scss
@@ -1,0 +1,1 @@
+@import "../../styles/Form.module.scss";
