+++ src/widgets/form/index.ts
@@ -1,0 +1,2 @@
+export { ControlledForm } from "./controlled-form/ui/ControlledForm";
+export { UncontrolledForm } from "./uncontrolled-form/ui/UncontrolledForm";
