+++ src/widgets/form/uncontrolled-form/ui/UncontrolledForm.module.scss
@@ -1,0 +1,1 @@
+@import "./../../styles/Form.module.scss";
