--- src/widgets/uncontrolled-form/index.ts
@@ -1,1 +1,0 @@
-export { UncontrolledForm } from "./ui/UncontrolledForm";
