--- src/widgets/uncontrolled-form/ui/UncontrolledForm.module.scss
@@ -1,82 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  max-width: 1200px;
-}
-
-.error {
-  position: absolute;
-  color: var(--error);
-  font-size: 12px;
-}
-
-.column {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  width: 100%;
-}
-
-.row {
-  display: flex;
-  gap: 18px;
-  width: 100%;
-}
-
-.control {
-  position: relative;
-  margin: 12px 0;
-
-  label {
-    display: inline-block;
-    margin-bottom: 5px;
-    margin-right: 5px;
-  }
-
-  input {
-    padding: 10px;
-    width: 100%;
-    box-sizing: border-box;
-  }
-
-  input[type="checkbox"] {
-    width: auto;
-    display: inline;
-  }
-
-  input[type="file"] {
-    padding: 0;
-    display: inline;
-    width: auto;
-  }
-
-  select {
-    padding: 10px;
-    width: 100%;
-    box-sizing: border-box;
-  }
-}
-
-.submit {
-  background-color: var(--background-secondary);
-  transition: all 0.2s ease;
-
-  &:hover {
-    background-color: var(--background-secondary-hover);
-  }
-
-  &:disabled {
-    cursor: not-allowed;
-    background-color: var(--background-secondary-hover);
-  }
-}
-
-.showPassword {
-  background-color: var(--background-secondary);
-  transition: all 0.2s ease;
-
-  &:hover {
-    background-color: var(--background-secondary-hover);
-  }
-}
